Added dynamic imports for 'useLayoutEffect' and moved TransitionLayout into separate file
