+++ hooks/animations/onload/useHeroTitleOnloadAnimations.js
@@ -1,4 +1,4 @@
-import { useLayoutEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { gsap } from "gsap/dist/gsap";
 import { useGSAP } from "@gsap/react";
 import { HERO as getHero } from "@/config/animations";
@@ -55,7 +55,7 @@ export const useHeroTitleOnloadAnimations = (titleRef, onLoadComplete) => {
   );
   const [transition, setTransition] = useState(null);
 
-  useLayoutEffect(() => {
+  useEffect(() => {
     if (!prefersReducedMotion) {
       setOpacity(0);
       setBlur(HERO.LOAD.TITLE.INITIAL_BLUR);
