+++ pages/_app.js
@@ -1,25 +1,52 @@
+// 1. Development imports
 if (process.env.NODE_ENV === "development") {
   require("../scripts/wdyr");
 }
 
+// 2. React and Next.js core imports
 import Head from "next/head";
+import dynamic from "next/dynamic";
+
+// 3. Styles
 import "@/styles/globals.css";
 import "@/styles/design-system.css";
 import "@/styles/mouse-follower.css";
+
+// 4. Configuration/Constants
 import { fonts } from "@/config/fonts";
-import Preloader from "@/components/Onload/Preloader";
-import Lines from "@/components/Onload/Lines";
-import { ScrollbarProvider } from "@/context/ScrollbarContext";
-import {
-  TransitionProvider,
-  TransitionLayout,
-} from "@/context/TransitionContext";
-import { PreloaderProvider } from "@/context/PreloaderContext";
+
+// 5. Context providers
 import { DimensionsProvider } from "@/context/DimensionsContext";
+import { PreloaderProvider } from "@/context/PreloaderContext";
+import { ScrollbarProvider } from "@/context/ScrollbarContext";
+import { TransitionProvider } from "@/context/TransitionContext";
+const ReducedMotionProvider = dynamic(
+  () =>
+    import("@/context/ReducedMotionContext").then(
+      (mod) => mod.ReducedMotionProvider
+    ),
+  { ssr: false }
+);
+
+// 6. Dynamic Components
+const Lines = dynamic(() => import("@/components/Onload/Lines"), {
+  ssr: false,
+});
+const CTA = dynamic(() => import("@/components/Global/CTA"), {
+  ssr: false,
+});
+const TransitionLayout = dynamic(
+  () =>
+    import("@/components/TransitionLayout").then((mod) => mod.TransitionLayout),
+  { ssr: false }
+);
+
+// 7. Components
+import Preloader from "@/components/Onload/Preloader";
 import CursorContainer from "@/components/Global/CursorContainer";
+
+// 8. Hooks
 import { usePreventStyleRemoval } from "@/hooks/navigation/usePreventStyleRemoval";
-import CTA from "@/components/Global/CTA";
-import { ReducedMotionProvider } from "@/context/ReducedMotionContext";
 
 // Persistent layer that stays mounted
 const PersistentLayer = () => {
